requirements.txt
alembic==1.9.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
Flask==2.2.2
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-Migrate==4.0.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
form==0.0.1
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.31
protobuf==3.20.1
psycopg2==2.9.5
pycparser==2.21
PyMySQL==1.0.2
SQLAlchemy==1.4.46
Werkzeug==2.2.2
WTForms==3.0.1
